Create memory_context.rs
